tailcat: keep lifecycle test with package tests
